+++ pages/_app.js
@@ -1,16 +1,16 @@
 import '../global.css'
 import "bootstrap/dist/css/bootstrap.css";
-import { ChakraProvider } from '@chakra-ui/react';
+// import Head from 'next/head';
 
+import CustomHead from '../components/CustomHead';
 
 export default function MyApp({ Component, pageProps }) {
-    // Use the layout defined at the page level, if available
-    const getLayout = Component.getLayout || ((page) => page)
 
-
-    return getLayout(
-        <ChakraProvider>
+    return (
+        <>
+            <CustomHead/>
             <Component{...pageProps} />,
-        </ChakraProvider>
+        </>
     )
+
 }
