--- variables.scss
@@ -1,4 +1,0 @@
-$carousel-control-color : #03c4eb;
-
-@import "./node_modules/bootstrap/scss/bootstrap";
-
